remove onlypasswordform
