--- src/pages/SignUp.tsx
@@ -1,11 +1,0 @@
-import Signup from "components/Signup";
-
-export default function SignUp() {
-  return (
-    <div className="grid place-items-center py-14">
-      <div className="w-full max-w-md px-7 py-8 bg-white border border-gray-l4 rounded-2xl">
-        <Signup />
-      </div>
-    </div>
-  );
-}
