+++ src/pages/SignUp/types.ts
@@ -1,5 +1,3 @@
-export type Donor = { email: string; firstName: string; lastName: string };
-
 export type CodeRecipientEmail = {
   raw: string;
   obscured: string;
